py_mrp_functions/requirements.txt
greenlet==1.1.2
numpy==1.19.5
pandas==1.1.5
PyMySQL==1.0.2
python-dateutil==2.8.2
pytz==2022.1
six==1.16.0
SQLAlchemy==1.4.39

Flask~=2.0.3
waitress~=2.0.0
celery~=5.1.2
click~=7.1
redis~=4.3.0
flower~=1.2.0
python-dotenv~=0.20.0
cryptography
